fix: move add/remove book logic to book service and adjust routes
- Moved the logic for adding and removing books from lists to book service
- Updated addBookToList and removeBookFromList methods to handle book and list IDs in the URL
