+++ src/modules/book/book.controller.ts
@@ -1,7 +1,8 @@
-import { Controller, Get, Param, Req, UseGuards } from "@nestjs/common";
+import { Body, Controller, Delete, Get, Param, Post, Req, UseGuards } from "@nestjs/common";
 import { BookService } from "./book.service";
 import { JwtAuthGuard } from "../auth/auth.guard";
 import { Request } from "express";
+import { AddBookToListDto } from "./dto/addBooktoList.dto";
 
 @Controller('app/book')
 export class BookController {
@@ -29,8 +30,22 @@ export class BookController {
 
   @UseGuards(JwtAuthGuard)
   @Get('list/:listId')
-  async getBooksByListId(@Param('listId') listId: string, @Req() req: Request) {
+  async getBooksByList(@Param('listId') listId: string, @Req() req: Request) {
     const requestedUserId = req.user['userId'];
-    return this.bookService.getBooksByListId(listId, requestedUserId);
+    return this.bookService.getBooksByList(listId, requestedUserId);
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Post(":bookId/:listId")
+  async addBookToList(@Param('bookId') bookId: string, @Param('listId') listId: string, @Body() data: AddBookToListDto, @Req() req: Request) {
+    const requestedUserId = req.user['userId'];
+    return this.bookService.addBookToList(bookId, listId, requestedUserId, data);
+  }
+
+  @UseGuards(JwtAuthGuard)
+  @Delete(":bookId/:listId")
+  async removeBookFromList(@Param('listId') listId: string, @Param('bookId') bookId: string, @Req() req: Request) {
+    const requestedUserId = req.user['userId'];
+    return this.bookService.removeBookFromList(listId, requestedUserId, bookId);
   }
 }
