+++ src/modules/book/book.module.ts
@@ -1,4 +1,4 @@
-import { forwardRef, Module } from "@nestjs/common";
+import { Module } from "@nestjs/common";
 import { PrismaService } from "../../prisma.service";
 import { BookService } from "./book.service";
 import { BookController } from "./book.controller";
@@ -8,9 +8,11 @@ import { GenreModule } from "../genre/genre.module";
 import { AuthorModule } from "../author/author.module";
 import { PublisherModule } from "../publisher/publisher.module";
 import { ListModule } from "../list/list.module";
+import { StatusModule } from "../status/status.module";
+import { AuthorizationModule } from "../authorization/authorization.module";
 
 @Module({
-  imports: [HttpModule, GenreModule, AuthorModule, PublisherModule, forwardRef(() => ListModule)],
+  imports: [HttpModule, GenreModule, AuthorModule, PublisherModule, ListModule, StatusModule, AuthorizationModule],
   controllers: [BookController],
   providers: [PrismaService, ExternalBookService, BookService],
   exports: [BookService],
